+++ configs/semi_det/_base_/coco_detection_voc.yml
@@ -1,6 +1,6 @@
 metric: COCO
 num_classes: 20
-#before train, change voc to coco with voc2coco.py
+# before training, change VOC to COCO format by 'python voc2coco.py'
 # partial labeled COCO, use `SemiCOCODataSet` rather than `COCODataSet`
 TrainDataset:
   !SemiCOCODataSet
@@ -22,5 +22,10 @@ EvalDataset:
   !COCODataSet
     image_dir: VOC2007/JPEGImages
     anno_path: PseudoAnnotations/VOC2007_test.json
-    dataset_dir:  dataset/voc/VOCdevkit/
+    dataset_dir: dataset/voc/VOCdevkit/
     allow_empty: true
+
+TestDataset:
+  !ImageFolder
+    anno_path: PseudoAnnotations/VOC2007_test.json # also support txt (like VOC's label_list.txt)
+    dataset_dir: dataset/voc/VOCdevkit/ # if set, anno_path will be 'dataset_dir/anno_path'
