+++ configs/semi_det/arsl/_base_/arsl_fcos_reader.yml
@@ -1,4 +1,4 @@
-worker_num: 1
+worker_num: 2
 SemiTrainReader:
   sample_transforms:
     - Decode: {}
@@ -35,22 +35,21 @@ SemiTrainReader:
 
 EvalReader:
   sample_transforms:
-  - Decode: {}
-  - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
-  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - Permute: {}
+    - Decode: {}
+    - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
+    - Permute: {}
   batch_transforms:
-  - PadBatch: {pad_to_stride: 32}
+    - PadBatch: {pad_to_stride: 32}
   batch_size: 1
-  shuffle: false
+
 
 TestReader:
   sample_transforms:
-  - Decode: {}
-  - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
-  - NormalizeImage: {is_scale: true, mean: [0.485,0.456,0.406], std: [0.229, 0.224,0.225]}
-  - Permute: {}
+    - Decode: {}
+    - Resize: {interp: 2, target_size: [800, 1333], keep_ratio: True}
+    - NormalizeImage: {mean: [0.485, 0.456, 0.406], std: [0.229, 0.224, 0.225], is_scale: true}
+    - Permute: {}
   batch_transforms:
-  - PadBatch: {pad_to_stride: 32}
+    - PadBatch: {pad_to_stride: 32}
   batch_size: 1
-  shuffle: false
