+++ configs/semi_det/arsl/_base_/optimizer_360k.yml
@@ -1,6 +1,6 @@
 epoch: 120 # employ iter to control shedule
 LearningRate:
-  base_lr: 0.02 # 0.02 for 64 batch
+  base_lr: 0.02 # 0.02 for 8*(4+4) batch
   schedulers:
   - !PiecewiseDecay
     gamma: 0.1
@@ -9,7 +9,7 @@ LearningRate:
     start_factor: 0.3333333333333333
     steps: 1000
 
-max_iter: 360000 # 360k for 32 batch, 720k for 16b tach
+max_iter: 360000 # 360k for 32 batch, 720k for 16 batch
 epoch_iter: 3000 # set epoch_iter for saving checkpoint and eval
 optimize_rate: 1
 SEMISUPNET:
@@ -17,8 +17,7 @@ SEMISUPNET:
   TEACHER_UPDATE_ITER: 1
   BURN_UP_STEP: -1
   EMA_KEEP_RATE: 0.9996
-  UNSUP_LOSS_WEIGHT: 1.0
-  # 20220817
+  UNSUP_LOSS_WEIGHT: 1.0 # detailed weights for cls and loc task can be seen in cr_loss
   PSEUDO_WARM_UP_STEPS: 2000
 
 OptimizerBuilder:
