+++ ppdet/modeling/heads/fcos_head.py
@@ -26,7 +26,7 @@ from paddle.nn.initializer import Normal, Constant
 from ppdet.core.workspace import register
 from ppdet.modeling.layers import ConvNormLayer, MultiClassNMS
 
-__all__ = ['FCOSFeat', 'FCOSHead', 'FCOSHead_SSOD']
+__all__ = ['FCOSFeat', 'FCOSHead', 'FCOSHead_ARSL']
 
 
 class ScaleReg(nn.Layer):
@@ -377,9 +377,9 @@ class FCOSHead(nn.Layer):
 
 
 @register
-class FCOSHead_SSOD(nn.Layer):
+class FCOSHead_ARSL(FCOSHead):
     """
-    FCOSHead
+    FCOSHead of ARSL for semi-det(ssod)
     Args:
         fcos_feat (object): Instance of 'FCOSFeat'
         num_classes (int): Number of classes
@@ -388,21 +388,26 @@ class FCOSHead_SSOD(nn.Layer):
         fcos_loss (object): Instance of 'FCOSLoss'
         norm_reg_targets (bool): Normalization the regression target if true
         centerness_on_reg (bool): The prediction of centerness on regression or clssification branch
+        nms (object): Instance of 'MultiClassNMS'
+        trt (bool): Whether to use trt in nms of deploy
     """
     __inject__ = ['fcos_feat', 'fcos_loss', 'nms']
     __shared__ = ['num_classes', 'trt']
 
     def __init__(self,
-                 fcos_feat,
                  num_classes=80,
+                 fcos_feat='FCOSFeat',
                  fpn_stride=[8, 16, 32, 64, 128],
                  prior_prob=0.01,
-                 fcos_loss='FCOSLoss',
+                 multiply_strides_reg_targets=False,
                  norm_reg_targets=True,
                  centerness_on_reg=True,
+                 num_shift=0.5,
+                 sqrt_score=False,
+                 fcos_loss='FCOSLossMILC',
                  nms='MultiClassNMS',
                  trt=False):
-        super(FCOSHead_SSOD, self).__init__()
+        super(FCOSHead_ARSL, self).__init__()
         self.fcos_feat = fcos_feat
         self.num_classes = num_classes
         self.fpn_stride = fpn_stride
@@ -410,9 +415,13 @@ class FCOSHead_SSOD(nn.Layer):
         self.fcos_loss = fcos_loss
         self.norm_reg_targets = norm_reg_targets
         self.centerness_on_reg = centerness_on_reg
+        self.multiply_strides_reg_targets = multiply_strides_reg_targets
+        self.num_shift = num_shift
         self.nms = nms
         if isinstance(self.nms, MultiClassNMS) and trt:
             self.nms.trt = trt
+        self.sqrt_score = sqrt_score
+
         conv_cls_name = "fcos_head_cls"
         bias_init_value = -math.log((1 - self.prior_prob) / self.prior_prob)
         self.fcos_head_cls = self.add_sublayer(
@@ -461,34 +470,7 @@ class FCOSHead_SSOD(nn.Layer):
             scale_reg = self.add_sublayer(feat_name, ScaleReg())
             self.scales_regs.append(scale_reg)
 
-    def _compute_locations_by_level(self, fpn_stride, feature):
-        """
-        Compute locations of anchor points of each FPN layer
-        Args:
-            fpn_stride (int): The stride of current FPN feature map
-            feature (Tensor): Tensor of current FPN feature map
-        Return:
-            Anchor points locations of current FPN feature map
-        """
-        shape_fm = paddle.shape(feature)
-        shape_fm.stop_gradient = True
-        h, w = shape_fm[2], shape_fm[3]
-        shift_x = paddle.arange(0, w * fpn_stride, fpn_stride)
-        shift_y = paddle.arange(0, h * fpn_stride, fpn_stride)
-        shift_x = paddle.unsqueeze(shift_x, axis=0)
-        shift_y = paddle.unsqueeze(shift_y, axis=1)
-        shift_x = paddle.expand(shift_x, shape=[h, w])
-        shift_y = paddle.expand(shift_y, shape=[h, w])
-        shift_x.stop_gradient = True
-        shift_y.stop_gradient = True
-        shift_x = paddle.reshape(shift_x, shape=[-1])
-        shift_y = paddle.reshape(shift_y, shape=[-1])
-        location = paddle.stack(
-            [shift_x, shift_y], axis=-1) + float(fpn_stride) / 2
-        location.stop_gradient = True
-        return location
-
-    def forward(self, fpn_feats):
+    def forward(self, fpn_feats, targets=None):
         assert len(fpn_feats) == len(
             self.fpn_stride
         ), "The size of fpn_feats is not equal to size of fpn_stride"
@@ -524,55 +506,55 @@ class FCOSHead_SSOD(nn.Layer):
         else:
             return cls_logits_list, bboxes_reg_list, centerness_list
 
-    def get_loss(self, fcos_head_outs, tag_labels, tag_bboxes, tag_centerness):
-        cls_logits, bboxes_reg, centerness = fcos_head_outs
-        return self.fcos_loss(cls_logits, bboxes_reg, centerness, tag_labels,
-                              tag_bboxes, tag_centerness)
+    # def get_loss(self, fcos_head_outs, tag_labels, tag_bboxes, tag_centerness):
+    #     cls_logits, bboxes_reg, centerness = fcos_head_outs
+    #     return self.fcos_loss(cls_logits, bboxes_reg, centerness, tag_labels,
+    #                           tag_bboxes, tag_centerness)
 
-    def post_process(self, fcos_head_outs, scale_factor):
-        locations, cls_logits, bboxes_reg, centerness = fcos_head_outs
-        pred_bboxes, pred_scores = [], []
+    # def post_process(self, fcos_head_outs, scale_factor):
+    #     locations, cls_logits, bboxes_reg, centerness = fcos_head_outs
+    #     pred_bboxes, pred_scores = [], []
 
-        for pts, cls, reg, ctn in zip(locations, cls_logits, bboxes_reg,
-                                      centerness):
-            scores, boxes = self._post_process_by_level(pts, cls, reg, ctn)
+    #     for pts, cls, reg, ctn in zip(locations, cls_logits, bboxes_reg,
+    #                                   centerness):
+    #         scores, boxes = self._post_process_by_level(pts, cls, reg, ctn)
 
-            pred_scores.append(scores)
-            pred_bboxes.append(boxes)
-        pred_bboxes = paddle.concat(pred_bboxes, axis=1)
-        pred_scores = paddle.concat(pred_scores, axis=1)
+    #         pred_scores.append(scores)
+    #         pred_bboxes.append(boxes)
+    #     pred_bboxes = paddle.concat(pred_bboxes, axis=1)
+    #     pred_scores = paddle.concat(pred_scores, axis=1)
 
-        # scale bbox to origin
-        scale_y, scale_x = paddle.split(scale_factor, 2, axis=-1)
-        scale_factor = paddle.concat(
-            [scale_x, scale_y, scale_x, scale_y], axis=-1).reshape([-1, 1, 4])
-        pred_bboxes /= scale_factor
+    #     # scale bbox to origin
+    #     scale_y, scale_x = paddle.split(scale_factor, 2, axis=-1)
+    #     scale_factor = paddle.concat(
+    #         [scale_x, scale_y, scale_x, scale_y], axis=-1).reshape([-1, 1, 4])
+    #     pred_bboxes /= scale_factor
 
-        pred_scores = pred_scores.transpose([0, 2, 1])
-        bbox_pred, bbox_num, _ = self.nms(pred_bboxes, pred_scores)
-        return bbox_pred, bbox_num
+    #     pred_scores = pred_scores.transpose([0, 2, 1])
+    #     bbox_pred, bbox_num, _ = self.nms(pred_bboxes, pred_scores)
+    #     return bbox_pred, bbox_num
 
-    def _post_process_by_level(self,
-                               locations,
-                               box_cls,
-                               box_reg,
-                               box_ctn,
-                               sqrt_score=False):
-        box_scores = F.sigmoid(box_cls).flatten(2).transpose([0, 2, 1])
-        box_centerness = F.sigmoid(box_ctn).flatten(2).transpose([0, 2, 1])
-        pred_scores = box_scores * box_centerness
-        if sqrt_score:
-            pred_scores = paddle.sqrt(pred_scores)
+    # def _post_process_by_level(self,
+    #                            locations,
+    #                            box_cls,
+    #                            box_reg,
+    #                            box_ctn,
+    #                            sqrt_score=False):
+    #     box_scores = F.sigmoid(box_cls).flatten(2).transpose([0, 2, 1])
+    #     box_centerness = F.sigmoid(box_ctn).flatten(2).transpose([0, 2, 1])
+    #     pred_scores = box_scores * box_centerness
+    #     if sqrt_score:
+    #         pred_scores = paddle.sqrt(pred_scores)
 
-        box_reg_ch_last = box_reg.flatten(2).transpose([0, 2, 1])
-        box_reg_decoding = paddle.stack(
-            [
-                locations[:, 0] - box_reg_ch_last[:, :, 0],
-                locations[:, 1] - box_reg_ch_last[:, :, 1],
-                locations[:, 0] + box_reg_ch_last[:, :, 2],
-                locations[:, 1] + box_reg_ch_last[:, :, 3]
-            ],
-            axis=1)
-        pred_boxes = box_reg_decoding.transpose([0, 2, 1])
+    #     box_reg_ch_last = box_reg.flatten(2).transpose([0, 2, 1])
+    #     box_reg_decoding = paddle.stack(
+    #         [
+    #             locations[:, 0] - box_reg_ch_last[:, :, 0],
+    #             locations[:, 1] - box_reg_ch_last[:, :, 1],
+    #             locations[:, 0] + box_reg_ch_last[:, :, 2],
+    #             locations[:, 1] + box_reg_ch_last[:, :, 3]
+    #         ],
+    #         axis=1)
+    #     pred_boxes = box_reg_decoding.transpose([0, 2, 1])
 
-        return pred_scores, pred_boxes
+    #     return pred_scores, pred_boxes
